refactor(utils): privatize utils module
BREAKING CHANGE:
utils module removed
